Added scroll pane and broke some other things
The scroll pane now exists. It can be used to select which files to be in an encounter.

Running into some issues with the files not being moved on Windows, though the changes are reflected in the program.

Also running into an issue where the program doesn't detect the changes in the folders while running.